Sign-Language-Translator-main 2/requirements.txt
Flask
opencv-python-headless
numpy
Pillow
mediapipe
tensorflow
language-tool-python